+++ tests/acceptance/mixed_driver_management/__init__.py
@@ -9,3 +9,5 @@ todo:
     when session ends and browser.config.hold_browser_open = False
     (currently we schedule automatic closure only for drivers spawned by selene)
 """
+
+# TODO: make these tests to live inside one py module with different test classes