+++ tests/base_test.py
@@ -37,3 +37,4 @@ class BaseTest:
 
     def teardown_method(self):
         browser.quit()
+        browser.config.driver = ...